pre color2/piece6 to piece12 change

+++ perft.cpp
@@ -97,12 +97,14 @@ int64_t perft(Position& pos, int depth, bool divide, int64_t& illegal_moves)
         return 1;
 
     int64_t legal_moves = 0;
-    int total_moves = 0;
+    int64_t total_moves = 0;
     char pp[6] = { '!', 'n', 'b', 'r', 'q', '!' };
 
     Gen::Move moves[256];
     Gen::Move* m = Gen::add_pseudo_moves(moves, pos);
     total_moves = m - moves;
+
+    //if (depth == 1) return total_moves;
 
     for (int i = 0; i < total_moves; i++) {
         Gen::Undo undo;
